Move logic to shared hook
